fix forecast model variable and remove logs

+++ RxJavaExample/app/src/main/java/com/example/germangirod/rxjavaexample/api/model/Rain.java
@@ -7,14 +7,13 @@ import com.google.gson.annotations.SerializedName;
  */
 public class Rain {
 
-    @SerializedName("3h")
-    private int value;
+    @SerializedName("3h") private float value;
 
-    public int getValue() {
+    public float getValue() {
         return value;
     }
 
-    public void setValue(int value) {
+    public void setValue(float value) {
         this.value = value;
     }
 }
